+++ acacore/reference_files/__init__.py
@@ -1,2 +1,2 @@
-from .ref_files import get_actions
-from .ref_files import get_custom_signatures
+from .get import get_actions
+from .get import get_custom_signatures
